+++ jabgui/src/test/java/org/jabref/gui/fieldeditors/optioneditors/LanguageEditorViewModelTest.java
@@ -8,7 +8,7 @@ import javax.swing.undo.UndoManager;
 
 import org.jabref.gui.autocompleter.SuggestionProvider;
 import org.jabref.logic.FilePreferences;
-import org.jabref.logic.citationkeypattern.CitationKeyGenerator;
+import org.jabref.logic.citationkeypattern.CitationKeyPatternPreferences;
 import org.jabref.logic.integrity.FieldCheckers;
 import org.jabref.logic.journals.JournalAbbreviationRepository;
 import org.jabref.model.database.BibDatabaseContext;
@@ -32,7 +32,7 @@ public class LanguageEditorViewModelTest {
         BibDatabaseContext databaseContext = mock(BibDatabaseContext.class);
         FilePreferences filePreferences = mock(FilePreferences.class);
         JournalAbbreviationRepository abbreviationRepository = mock(JournalAbbreviationRepository.class);
-        FieldCheckers fieldCheckers = new FieldCheckers(databaseContext, filePreferences, abbreviationRepository, false, CitationKeyGenerator.DEFAULT_UNWANTED_CHARACTERS);
+        FieldCheckers fieldCheckers = new FieldCheckers(databaseContext, filePreferences, abbreviationRepository, false, CitationKeyPatternPreferences.DEFAULT_UNWANTED_CHARACTERS);
         SuggestionProvider<?> suggestionProvider = mock(SuggestionProvider.class);
 
         languageEditorViewModel = new LanguageEditorViewModel(
