+++ jablib/src/main/java/org/jabref/logic/citationkeypattern/CitationKeyGenerator.java
@@ -25,15 +25,6 @@ public class CitationKeyGenerator extends BracketedPattern {
     /// All single characters that we can use for extending a key to make it unique.
     public static final String APPENDIX_CHARACTERS = "abcdefghijklmnopqrstuvwxyz";
 
-    /// List of unwanted characters. These will be removed at the end.
-    /// Note that `+` is a wanted character to indicate "et al." in authorsAlpha.
-    /// Example: `ABC+`. See `org.jabref.logic.citationkeypattern.BracketedPatternTest#authorsAlpha()` for examples.
-    ///
-    /// Should never be used directly - use [CitationKeyPatternPreferences#getUnwantedCharacters()] instead, which respects the user's preferences
-    ///
-    /// See also #DISALLOWED_CHARACTERS
-    public static final String DEFAULT_UNWANTED_CHARACTERS = "-`ʹ:!;?^$";
-
     /// Source of disallowed characters: <https://tex.stackexchange.com/a/408548/9075>
     /// These characters are disallowed in BibTeX keys.
     public static final List<Character> DISALLOWED_CHARACTERS = Arrays.asList('{', '}', '(', ')', ',', '=', '\\', '"', '#', '%', '~', '\'');
