+++ jablib/src/test/java/org/jabref/logic/citationkeypattern/CitationKeyGeneratorTest.java
@@ -20,7 +20,7 @@ import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.MethodSource;
 import org.mockito.Answers;
 
-import static org.jabref.logic.citationkeypattern.CitationKeyGenerator.DEFAULT_UNWANTED_CHARACTERS;
+import static org.jabref.logic.citationkeypattern.CitationKeyPatternPreferences.DEFAULT_UNWANTED_CHARACTERS;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.mock;
@@ -274,17 +274,17 @@ class CitationKeyGeneratorTest {
 
     @Test
     void removeIllegalCharactersPreservesDiacritics() {
-        assertEquals("kṛṣṇā", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("kṛṣṇā", CitationKeyGenerator.DEFAULT_UNWANTED_CHARACTERS));
-        assertEquals("Müller2020", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("Müller2020", CitationKeyGenerator.DEFAULT_UNWANTED_CHARACTERS));
-        assertEquals("taṇḍulīyaka", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("taṇḍulīyaka", CitationKeyGenerator.DEFAULT_UNWANTED_CHARACTERS));
-        assertEquals("García_2021", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("García_2021", CitationKeyGenerator.DEFAULT_UNWANTED_CHARACTERS));
+        assertEquals("kṛṣṇā", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("kṛṣṇā", CitationKeyPatternPreferences.DEFAULT_UNWANTED_CHARACTERS));
+        assertEquals("Müller2020", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("Müller2020", CitationKeyPatternPreferences.DEFAULT_UNWANTED_CHARACTERS));
+        assertEquals("taṇḍulīyaka", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("taṇḍulīyaka", CitationKeyPatternPreferences.DEFAULT_UNWANTED_CHARACTERS));
+        assertEquals("García_2021", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("García_2021", CitationKeyPatternPreferences.DEFAULT_UNWANTED_CHARACTERS));
     }
 
     @Test
     void removeIllegalCharactersRemovesDisallowedChars() {
-        assertEquals("key", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("ke{y}", CitationKeyGenerator.DEFAULT_UNWANTED_CHARACTERS));
-        assertEquals("key", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("ke y", CitationKeyGenerator.DEFAULT_UNWANTED_CHARACTERS));
-        assertEquals("key", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("k#e%y", CitationKeyGenerator.DEFAULT_UNWANTED_CHARACTERS));
+        assertEquals("key", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("ke{y}", CitationKeyPatternPreferences.DEFAULT_UNWANTED_CHARACTERS));
+        assertEquals("key", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("ke y", CitationKeyPatternPreferences.DEFAULT_UNWANTED_CHARACTERS));
+        assertEquals("key", CitationKeyGenerator.removeUnwantedCharactersWithKeepDiacritics("k#e%y", CitationKeyPatternPreferences.DEFAULT_UNWANTED_CHARACTERS));
     }
 
     @ParameterizedTest
